chore: cleanup types
